+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG.md
 
+## 0.7.0 (2024-01-12)
+
+- Store un-hexed multihash in Dict4.hash
+- Use Dict4.V2 to add that hash to the struct
+- Drop version down to "v2" (from "v4")
+
+## 0.6.2 (2023-10-22)
+
+- Cleanup for demo
+
 ## 0.5.1 (2023-10-21)
 
 - Remove legacy code